Lab10/requirements.txt
psycopg2-binary
pygame
